Use camelCase and TitleCase for caches, plugins, etc filenames
