+++ lib/ArchetypeSystem.js
