+++ lib/AssetCache.js
