+++ lib/CacheSerializerFactory.js
@@ -74,8 +74,8 @@
  * @returns {Promise} promise that resolves when writing completes
  */
 
-const FileSerializerPlugin = require('./hard-source-file-serializer-plugin');
-const Append2SerializerPlugin = require('./hard-source-append-2-serializer-plugin');
+const FileSerializerPlugin = require('./HardSourceFileSerializerPlugin');
+const Append2SerializerPlugin = require('./HardSourceAppend2SerializerPlugin');
 
 const pluginCompat = require('./util/plugin-compat');
