+++ lib/EnhancedResolveCache.js
