+++ lib/HardAssetPlugin.js
