+++ lib/HardBasicDependencyPlugin.js
@@ -29,7 +29,7 @@ class HardBasicDependencyPlugin {
 
   apply(compiler) {
     if (this.options.schema < 4) {
-      const HardBasicDependencyPluginLegacy = require('./hard-basic-dependency-plugin-legacy');
+      const HardBasicDependencyPluginLegacy = require('./HardBasicDependencyPluginLegacy');
       new HardBasicDependencyPluginLegacy(this.options).apply(compiler);
     } else {
       const schemas = require('./schema-4');
