+++ lib/HardBasicDependencyPluginLegacy.js
@@ -1,5 +1,5 @@
 const cachePrefix = require('./util').cachePrefix;
-const LoggerFactory = require('./logger-factory');
+const LoggerFactory = require('./loggerFactory');
 const pluginCompat = require('./util/plugin-compat');
 const relateContext = require('./util/relate-context');
