+++ lib/HardCompilationPlugin.js
