+++ lib/HardConcatenationModulePlugin.js
