+++ lib/HardDependencyBlockPlugin.js
