+++ lib/HardGeneratorPlugin.js
