+++ lib/HardModuleAssetsPlugin.js
