+++ lib/HardModuleErrorsPlugin.js
