+++ lib/HardModuleExtractTextPlugin.js
