+++ lib/HardModuleMiniCssExtractPlugin.js
