+++ lib/HardNormalModuleFactoryPlugin.js
