+++ lib/HardNormalModulePlugin.js
