+++ lib/HardParserPlugin.js
