+++ lib/HardSourceAppend2Serializer.js
