+++ lib/HardSourceAppend2SerializerPlugin.js
@@ -2,7 +2,7 @@ const join = require('path').join;
 
 const pluginCompat = require('./util/plugin-compat');
 
-let AppendSerializer;
+let Append2Serializer;
 
 const _blockSizeByName = {
   data: 4 * 1024,
@@ -13,15 +13,15 @@ const _blockSizeByName = {
   resolver: 256,
 };
 
-class HardSourceAppendSerializerPlugin {
+class HardSourceAppend2SerializerPlugin {
   apply(compiler) {
     pluginCompat.tap(
       compiler,
       'hardSourceCacheFactory',
-      'AppendSerializer',
+      'Append2Serializer',
       factory => info => {
         if (info.type === 'data') {
-          return HardSourceAppendSerializerPlugin.createSerializer(info);
+          return HardSourceAppend2SerializerPlugin.createSerializer(info);
         }
         return factory(info);
       },
@@ -29,20 +29,20 @@ class HardSourceAppendSerializerPlugin {
   }
 }
 
-HardSourceAppendSerializerPlugin.createSerializer = ({
+HardSourceAppend2SerializerPlugin.createSerializer = ({
   cacheDirPath,
   name,
   autoParse,
 }) => {
-  if (!AppendSerializer) {
-    AppendSerializer = require('./hard-source-append-serializer');
+  if (!Append2Serializer) {
+    Append2Serializer = require('./HardSourceAppend2Serializer');
   }
 
-  return new AppendSerializer({
+  return new Append2Serializer({
     cacheDirPath: join(cacheDirPath, name),
     blockSize: _blockSizeByName[name],
     autoParse: autoParse,
   });
 };
 
-module.exports = HardSourceAppendSerializerPlugin;
+module.exports = HardSourceAppend2SerializerPlugin;
