+++ lib/HardSourceAppendSerializer.js
