+++ lib/HardSourceAppendSerializerPlugin.js
@@ -2,7 +2,7 @@ const join = require('path').join;
 
 const pluginCompat = require('./util/plugin-compat');
 
-let Append2Serializer;
+let AppendSerializer;
 
 const _blockSizeByName = {
   data: 4 * 1024,
@@ -13,15 +13,15 @@ const _blockSizeByName = {
   resolver: 256,
 };
 
-class HardSourceAppend2SerializerPlugin {
+class HardSourceAppendSerializerPlugin {
   apply(compiler) {
     pluginCompat.tap(
       compiler,
       'hardSourceCacheFactory',
-      'Append2Serializer',
+      'AppendSerializer',
       factory => info => {
         if (info.type === 'data') {
-          return HardSourceAppend2SerializerPlugin.createSerializer(info);
+          return HardSourceAppendSerializerPlugin.createSerializer(info);
         }
         return factory(info);
       },
@@ -29,20 +29,20 @@ class HardSourceAppend2SerializerPlugin {
   }
 }
 
-HardSourceAppend2SerializerPlugin.createSerializer = ({
+HardSourceAppendSerializerPlugin.createSerializer = ({
   cacheDirPath,
   name,
   autoParse,
 }) => {
-  if (!Append2Serializer) {
-    Append2Serializer = require('./hard-source-append-2-serializer');
+  if (!AppendSerializer) {
+    AppendSerializer = require('./HardSourceAppendSerializer');
   }
 
-  return new Append2Serializer({
+  return new AppendSerializer({
     cacheDirPath: join(cacheDirPath, name),
     blockSize: _blockSizeByName[name],
     autoParse: autoParse,
   });
 };
 
-module.exports = HardSourceAppend2SerializerPlugin;
+module.exports = HardSourceAppendSerializerPlugin;
