+++ lib/HardSourceCacacheSerializer.js
