+++ lib/HardSourceCacacheSerializerPlugin.js
@@ -25,7 +25,7 @@ HardSourceCacacheSerializerPlugin.createSerializer = ({
   name,
 }) => {
   if (!CacacheSerializer) {
-    CacacheSerializer = require('./hard-source-cacache-serializer');
+    CacacheSerializer = require('./HardSourceCacacheSerializer');
   }
 
   return new CacacheSerializer({
