+++ lib/HardSourceFileSerializer.js
