+++ lib/HardSourceFileSerializerPlugin.js
@@ -22,7 +22,7 @@ class HardSourceFileSerializerPlugin {
 
 HardSourceFileSerializerPlugin.createSerializer = ({ cacheDirPath, name }) => {
   if (!FileSerializer) {
-    FileSerializer = require('./hard-source-file-serializer');
+    FileSerializer = require('./HardSourceFileSerializer');
   }
 
   return new FileSerializer({
