+++ lib/HardSourceJsonSerializer.js
