+++ lib/HardSourceJsonSerializerPlugin.js
@@ -22,7 +22,7 @@ class HardSourceJsonSerializerPlugin {
 
 HardSourceJsonSerializerPlugin.createSerializer = ({ cacheDirPath, name }) => {
   if (!JsonSerializer) {
-    JsonSerializer = require('./hard-source-json-serializer');
+    JsonSerializer = require('./HardSourceJsonSerializer');
   }
 
   return new JsonSerializer({
