+++ lib/HardSourceLeveldbSerializer.js
