+++ lib/HardSourceLeveldbSerializerPlugin.js
@@ -24,7 +24,7 @@ class HardSourceLevelDbSerializerPlugin {
 HardSourceLevelDbSerializerPlugin.createSerializer = info => {
   if (!LevelDbSerializer) {
     try {
-      LevelDbSerializer = require('./hard-source-leveldb-serializer');
+      LevelDbSerializer = require('./HardSourceLeveldbSerializer');
     } catch (e) {}
   }
 
@@ -34,7 +34,7 @@ HardSourceLevelDbSerializerPlugin.createSerializer = info => {
     });
   } else {
     if (!AppendSerializerPlugin) {
-      AppendSerializerPlugin = require('./hard-source-append-serializer-plugin');
+      AppendSerializerPlugin = require('./HardSourceAppendSerializerPlugin');
     }
 
     return AppendSerializerPlugin.createSerializer(info);
