+++ lib/HardSourceSourcePlugin.js
