+++ lib/Md5Cache.js
