+++ lib/ModuleCache.js
