+++ lib/ModuleResolverCache.js
