+++ lib/defaultConfigHash.js
