+++ lib/envHash.js
