+++ lib/loggerFactory.js
