+++ lib/util/index.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 
-const LoggerFactory = require('../logger-factory');
+const LoggerFactory = require('../loggerFactory');
 
 exports.cachePrefix = cachePrefix;
