+++ lib/util/log-messages.js
@@ -1,5 +1,5 @@
 const cachePrefix = require('.').cachePrefix;
-const LoggerFactory = require('../logger-factory');
+const LoggerFactory = require('../loggerFactory');
 
 exports.moduleFreezeError = (compilation, module, e) => {
   const loggerSerial = LoggerFactory.getLogger(compilation).from('serial');
